rename steps components
