+++ src/components/relaying-education-series/index.tsx
@@ -1,7 +1,7 @@
-import WhatIsRelaying from '@/components/relaying-education-series/steps/what-is-relaying'
-import HowItWorks from '@/components/relaying-education-series/steps/how-it-works'
-import Benefits from '@/components/relaying-education-series/steps/benefits'
-import TechnicalInformation from '@/components/relaying-education-series/steps/technical-information'
+import WhatIsRelaying from '@/components/relaying-education-series/steps/WhatIsRelaying'
+import HowItWorks from '@/components/relaying-education-series/steps/HowItWorks'
+import Benefits from '@/components/relaying-education-series/steps/Benefits'
+import TechnicalInformation from '@/components/relaying-education-series/steps/TechnicalInformation'
 import RelaySeriesStepper from '@/components/relaying-education-series/RelaySeriesStepper'
 import { type EducationSeriesStepperProps } from '@/components/relaying-education-series/RelaySeriesStepper/useEducationSeriesStepper'
 import { useRouter } from 'next/router'
