+++ src/components/relaying-education-series/steps/TechnicalInformation/index.tsx
